replace AVA with jest
