--- ava.config.cjs
@@ -1,12 +1,0 @@
-module.exports = {
-    files: [
-      "./test/*"
-    ],
-    extensions: {
-      "ts": "module"
-    },
-    nodeArguments: [
-      "--loader=ts-node/esm",
-      "--experimental-specifier-resolution=node"
-    ]
-}
