+++ package.json
@@ -7,7 +7,7 @@
   "license": "MIT",
   "scripts": {
     "lint": "eslint . --ext .js,.ts,.json",
-    "test": "yarn ava --verbose",
+    "test": "jest",
     "type-check": "tsc --noEmit",
     "type-check:watch": "yarn type-check -- --watch",
     "build:types": "tsc --declaration",
@@ -17,12 +17,11 @@
     "clean": "yarn del ./dist"
   },
   "devDependencies": {
-    "@ava/typescript": "^2.0.0",
     "@types/express": "^4.17.11",
+    "@types/jest": "^29.4.0",
     "@types/node": "^15.0.1",
     "@typescript-eslint/eslint-plugin": "^5.3.0",
     "@typescript-eslint/parser": "^5.3.0",
-    "ava": "^4.0.1",
     "del-cli": "^4.0.1",
     "esbuild": "^0.13.12",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -30,7 +29,10 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-json": "^3.1.0",
+    "jest": "^29.4.3",
+    "jest-each": "^29.4.3",
     "prettier": "^2.4.1",
+    "ts-jest": "^29.0.5",
     "ts-node": "^10.3.1",
     "typescript": "^4.2.4"
   },
